ai_hallucination/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# AI and LLM
google-generativeai>=0.3.0
tenacity>=8.2.0  # For exponential backoff retry logic

# Database (PostgreSQL Async)
sqlmodel>=0.0.8
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.11.0  # For database migrations

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0